Define THESIS_TYPES in ThesisLink model itself
This is recommended. Also give names like
MASTER_THESIS = 'MA' to avoid magic strings and allow for easy
reference, e.g. thesis.type = ThesisLink.MASTER_THESIS. This sort
of thing helps in maintainability of the codebase, since if we want
to change the abbreviation to 'MASTER' for some reason, we have it
defined in one place, and not in 100 places in the code.