Set blank=False on domain field of ThesisLink
A blank domain throws an error in /theses/.
I've also installed factory_boy (pip install -r requirements.txt
to get it) and set up a ThesisLinkFactory and a unit tests that
tests whether ThesisLink validates for a blank domain field.

Not sure if this is exactly how you're supposed to validate
a model, but it's a start. First unit test of the project, whoohoo!