Move as much logic as possible to submission model
The submission process contains some logic that
could be either handled using default values in the
model (where possible) or abstracted into model
methods. This has the great advantage that the
individual parts can be unit-tested (which is a lot
harder to do with code in the view).

For now there is still some commented-out code for
easy reference. This code will be removed as it is
replaced by the classes above it.