Extract ThesisLink#header_as_table to partial template
Markup logic should be in templates, not in model layer.

Partial template is named _header_as_table to signal
that it is ment to be included into other templates
(analogous to the private method naming convention.)
The full text of several of the thesis link's fields
(domain, subject_area, etc.) are fetched from constant dicts
in custom filters defined in theses.templatetags.theses_extras.