Rewrite prefill_using_DOI view
I'm not happy with this yet. The template requires all forms to be
instantiated at all times, while relying on javascript to hide them.
There should just be an if-statement checking if the form object exists.
That way, the error handling, which is now on the form level, will render
properly and everything is much shorter.