Unstable: New ArxivCaller disabled Submit manuscript page
The removeal of the old ArxivCaller made the submit Manuscript
view crash. At least 5 properties of the old caller was relied on
but not found on the new caller.

Also, (more) important are three checks on the submit manuscript
page which are missing now. These should be implemented into the
submission form back again.