Updated login() view
The login view is update such that the users will not get any
unexpected/unpleasant errors anymore if they don't get their
password right the first time.

Also, the redirect_url didn't work when it was given through
a GET request, which is used for example after a redirect
from the @login_required decorator.