Add login with email functionality
Now users will be able to log in with their email as well as their username.
I make it implicit; I do not tell the user this in the login form, as I'm
not sure if the non-uniqueness of the `email` field could cause any problems.

I now built a rude try-catch to ease the login process for the user instead.