refactor: 🔒️ always create new profile upon registration
Registration now always creates a unique profile. THis patches a potential loophole where one could create contributor accounts for unregistered profiles and gain access to them. Additionally, it fixes a fatal crash when such a contributor exists and the profile was being matched to both the old and new contributors.